Changed evaluation time limit.

+++ nudge_point.rb
@@ -12,7 +12,7 @@ class NudgePoint
   end
   
   def evaluate (outcome_data)
-    if Time.now.to_i > outcome_data.start_moment + Outcome::TIME_LIMIT
+    if Time.now.to_f > outcome_data.start_moment + Outcome::TIME_LIMIT
       n = outcome_data.points_evaluated
       raise NudgeError::TimeLimitExceeded, "the time limit was exceeded after evaluating #{n} points"
     end
